my_tfhub/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.2.0=eigen
absl-py=0.10.0=py37_0
aiohttp=3.6.2=py37he774522_0
astor=0.8.1=py37_0
async-timeout=3.0.1=py37_0
attrs=20.2.0=py_0
blas=1.0=mkl
blinker=1.4=py37_0
brotlipy=0.7.0=py37he774522_1000
ca-certificates=2020.7.22=0
cachetools=4.1.1=py_0
certifi=2020.6.20=py37_0
cffi=1.14.3=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
click=7.1.2=py_0
cryptography=3.1.1=py37h7a1dbc1_0
cycler=0.10.0=py37_0
flask=1.1.2=pypi_0
flask-cors=3.0.9=pypi_0
freetype=2.10.2=hd328e21_0
gast=0.2.2=py37_0
google-auth=1.22.1=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
grpcio=1.31.0=py37he7da953_0
h5py=2.10.0=py37h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=py_0
importlib-metadata=1.7.0=py37_0
intel-openmp=2020.2=254
itsdangerous=1.1.0=pypi_0
jinja2=2.11.2=pypi_0
joblib=0.17.0=py_0
jpeg=9b=hb83a4c4_2
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_1
kiwisolver=1.2.0=py37h74a9793_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.13.0=h200bbdf_0
libtiff=4.1.0=h56a325e_1
lz4-c=1.9.2=hf4a77e7_3
markdown=3.3=py37_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py37hba9282a_0
mkl=2020.2=256
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
multidict=4.7.6=py37he774522_1
numpy=1.19.1=py37h5510c5b_0
numpy-base=1.19.1=py37ha3acd2a_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
openssl=1.1.1h=he774522_0
opt_einsum=3.1.0=py_0
pandas=1.1.3=py37ha925a31_0
pillow=7.2.0=py37hcc1f983_0
pip=20.2.3=py37_0
protobuf=3.13.0=py37h33f27b4_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pyjwt=1.7.1=py37_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h6538335_2
pyreadline=2.1=py37_1
pysocks=1.7.1=py37_1
python=3.7.9=h60c2a47_0
python-dateutil=2.8.1=py_0
pytz=2020.1=py_0
qt=5.9.7=vc14h73c81de_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=py_0
rsa=4.6=py_0
scikit-learn=0.23.2=py37h47e9c7a_0
scipy=1.5.2=py37h9439919_0
setuptools=50.3.0=py37h9490d1a_1
sip=4.19.8=py37h6538335_0
six=1.15.0=py_0
sqlite=3.33.0=h2a8f88b_0
tensorboard=2.2.1=pyh532a8cf_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.1.0=eigen_py37hd727fc0_0
tensorflow-base=2.1.0=eigen_py37h49b2757_0
tensorflow-estimator=2.1.0=pyhd54b08b_0
tensorflow-hub=0.9.0=pypi_0
termcolor=1.1.0=py37_1
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
torchvision=0.5.0=pypi_0
tornado=6.0.4=py37he774522_1
urllib3=1.25.10=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
werkzeug=0.16.1=py_0
wheel=0.35.1=py_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
wrapt=1.12.1=py37he774522_1
xlrd=1.2.0=py37_0
xz=5.2.5=h62dcd97_0
yarl=1.6.0=py37he774522_0
zipp=3.3.0=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
